requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
aiosqlite==0.16.0
async-timeout==3.0.1
atlastk==0.12.2
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
iso8601==0.1.13
multidict==5.1.0
pathlib==1.0.1
payeer-api==0.1
PyPika==0.44.1
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
tortoise==0.1.1
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
yarl==1.6.3
